+++ lib/ig_markets/instrument.rb
@@ -16,6 +16,12 @@ module IGMarkets
       attribute :settlement_info
     end
 
+    # Contains details on the limited risk premium for an instrument. Returned by {#limited_risk_premium}.
+    class LimitedRiskPremium < Model
+      attribute :unit, Symbol, allowed_values: [:percentage, :points]
+      attribute :value, Float
+    end
+
     # Contains details on the margin deposit requirements for an instrument at a certain price band. Returned by
     # {#margin_deposit_bands}.
     class MarginDepositBand < Model
@@ -65,6 +71,7 @@ module IGMarkets
     attribute :expiry, Date, nil_if: %w(- DFB), format: ['%d-%b-%y', '%b-%y']
     attribute :expiry_details, ExpiryDetails
     attribute :force_open_allowed, Boolean
+    attribute :limited_risk_premium, LimitedRiskPremium
     attribute :lot_size, Float
     attribute :margin_deposit_bands, MarginDepositBand
     attribute :margin_factor, Float